Add a cache of host information to the client code.
The host cache maps server uuids to cached information about each
server. This includes the server's name and a list of connection info
data classes. Connection info is initially obtained from a url and can
be refreshed periodically. Each connection info for a host is tested
and the fastest is used. The host cache is not yet persisted to disk.

Rework Client APIs to refer to hosts by uuid and look up uuid in host
cache for connection information.

Add a resolver abstraction for looking up all matching addresses for a
hostname.

Add mockable clock and timesource abstractions to facilitate testing.