Add Linux spi and gpio drivers.
These use the JNI to call native code which uses spidev ioctls and
libgpiod to drive the built-in hardware in a linux system.

The native library is not built by gradle, but rather builds tar and
zip files containing the native code and a Makefile. After using these
to build the native library on the target machine, it may be placed in
the same directory as the server jar file.

Tweak SPI interface. Add a parameter to allow large SPI transactions
to be split into multiple smaller transactions if needed by the
driver. Add a UByteArray variant of spiWrite.