Add a server-side-events stream endpoint.
Add an /api/v1/events endpoint which uses SSE to send a stream of
events to a client.

The only event send currently is an update event when the state is
updated, with the event data being the same as the new update tag from the
status endpoint.

Additionally an empty event with a comment of "heartbeat" will be sent
every 30 seconds to keep the connection alive.

Add an update tag to the status endpoint which corresponds to the
update event data.

Add a way to run a real Netty server instance with a random listening
port. It apears that KTor's test server doesn't support SSE.